Ajax

+++ app/controllers/answers_controller.rb
@@ -1,31 +1,37 @@
 class AnswersController < ApplicationController
   before_action :authenticate_user!
 
-  expose :question
-  expose :answer, parent: :question
+  expose :question, -> { find_question }
+  expose :answer, build: -> { question.answers.build(answer_params) }
 
   def create
     answer.author = current_user
-    if answer.save
-      redirect_to question, notice: 'Your answer has been sent successfully.'
-    else
-      render 'questions/show'
-    end
+    answer.save
+  end
+
+  def update
+    answer.update(answer_params) if current_user.author_of?(answer)
   end
 
   def destroy
-    answer = Answer.find(params[:id])
-    if current_user.author_of?(answer)
-      answer.destroy
-      redirect_to answer.question, notice: 'Your answer successfully deleted.'
-    else
-      render 'questions/show'
-    end
+    answer.destroy if current_user.author_of?(answer)
+  end
+
+  def best
+    question.update_best_answer(answer) if current_user.author_of?(question)
   end
 
   private
 
   def answer_params
     params.require(:answer).permit(:body)
+  end
+
+  def find_question
+    if params[:question_id]
+      Question.find(params[:question_id])
+    else
+      answer.question
+    end
   end
 end
